requirements.txt
Flask
Flask-Cache
redis
Pillow
numpy
waitress